Lab1Week1 -  V1.0
I added in the instance of, to check to see if it is the expected type, at the end of the program.

I finally added, a static way of implementing studentID's, so whenever it is created, there will be a unique ID starting from 1000 and incrementing each time the object is created.